pkg/nvme: delete unused code
this code is discovery-service code.
it was wrongfully ported to the discovery-client repo when the projects
were splitted.
